+++ trasherbox-api/index.js
@@ -1,320 +1,19 @@
 const express = require("express");
 const cors = require("cors");
 const bodyParser = require("body-parser");
-const sqlite3 = require("sqlite3").verbose();
-const bcrypt = require("bcryptjs");
-const fetch = require("node-fetch");
-
 const app = express();
 const PORT = 3001;
+
+const db = require("./db/connection");
 
 app.use(cors());
 app.use(bodyParser.json());
 
-// Conexión a SQLite
-const db = new sqlite3.Database("./trasherbox.db", (err) => {
-  if (err) console.error("❌ Error conectando a SQLite:", err.message);
-  else console.log("✅ Conectado a trasherbox.db");
-});
-
-// Crear tabla orden si no existe
-db.run(`
-  CREATE TABLE IF NOT EXISTS orden (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    token TEXT,
-    estado TEXT,
-    monto INTEGER,
-    fecha TEXT
-  )
-`);
-
-// Middleware para verificar admin
-function verifyAdmin(req, res, next) {
-  const email = req.body.adminEmail || req.query.adminEmail;
-  if (!email) return res.status(400).json({ error: "Falta adminEmail" });
-
-  db.get("SELECT rol FROM usuario WHERE email = ?", [email], (err, row) => {
-    if (err || !row) return res.status(403).json({ error: "Acceso denegado" });
-    if (row.rol !== "admin") return res.status(403).json({ error: "Solo administradores" });
-    next();
-  });
-}
-
-// Registro de usuario
-app.post("/api/register", async (req, res) => {
-  const { email, password, usuario, telefono } = req.body;
-  if (!email || !password || !usuario || !telefono)
-    return res.status(400).json({ error: "Faltan campos obligatorios" });
-
-  try {
-    const hashedPassword = await bcrypt.hash(password, 10);
-    db.run(
-      `INSERT INTO usuario (email, password_hash, usuario, telefono) VALUES (?, ?, ?, ?)`,
-      [email, hashedPassword, usuario, telefono],
-      function (err) {
-        if (err) return res.status(400).json({ error: "Usuario ya existe o error al registrar" });
-        res.status(200).json({ success: true, userId: this.lastID });
-      }
-    );
-  } catch {
-    res.status(500).json({ error: "Error del servidor" });
-  }
-});
-
-// Login
-app.post("/api/login", (req, res) => {
-  const { email, password } = req.body;
-  db.get(`SELECT * FROM usuario WHERE email = ?`, [email], async (err, row) => {
-    if (err || !row) return res.status(401).json({ error: "Credenciales incorrectas" });
-
-    const valid = await bcrypt.compare(password, row.password_hash);
-    if (!valid) return res.status(401).json({ error: "Credenciales incorrectas" });
-
-    res.json({
-      success: true,
-      user: {
-        id: row.id,
-        email: row.email,
-        usuario: row.usuario,
-        rol: row.rol,
-      },
-    });
-  });
-});
-
-// Ruta protegida de prueba
-app.post("/api/admin/secure", verifyAdmin, (req, res) => {
-  res.json({ message: "Ruta protegida para administradores" });
-});
-
-// Dashboard resumen
-app.get("/api/admin/dashboard-summary", (req, res) => {
-  const summary = {};
-  db.get("SELECT COUNT(*) AS totalUsuarios FROM usuario", (err, usuariosRow) => {
-    if (err) return res.status(500).json({ error: "Error al contar usuarios" });
-    summary.usuarios = usuariosRow.totalUsuarios;
-
-    db.get("SELECT COUNT(*) AS totalPedidos FROM orden", (err, pedidosRow) => {
-      if (err) return res.status(500).json({ error: "Error al contar pedidos" });
-      summary.ordenes = pedidosRow.totalPedidos;
-
-      db.get("SELECT COUNT(*) AS totalProductos FROM producto", (err, productosRow) => {
-        if (err) return res.status(500).json({ error: "Error al contar productos" });
-        summary.productos = productosRow.totalProductos;
-
-        db.get("SELECT SUM(monto) AS totalVentas FROM orden", (err, ventasRow) => {
-          if (err) return res.status(500).json({ error: "Error al calcular ventas" });
-          summary.ventas = ventasRow.totalVentas || 0;
-          res.json(summary);
-        });
-      });
-    });
-  });
-});
-
-// Productos por categoría
-app.get("/api/admin/productos-por-categoria", (req, res) => {
-  db.all(
-    `SELECT categoria, COUNT(*) as total 
-     FROM producto 
-     WHERE categoria IS NOT NULL AND categoria != '' 
-     GROUP BY categoria`,
-    (err, rows) => {
-      if (err) return res.status(500).json({ error: "Error al contar por categoría" });
-      res.json(rows);
-    }
-  );
-});
-
-// Listar usuarios
-app.post("/api/admin/users/list", verifyAdmin, (req, res) => {
-  db.all("SELECT id, email, usuario, telefono, rol FROM usuario", (err, rows) => {
-    if (err) return res.status(500).json({ error: "Error al listar usuarios" });
-    res.json(rows);
-  });
-});
-
-// Crear usuario
-app.post("/api/admin/users", verifyAdmin, async (req, res) => {
-  const { email, password, usuario, telefono, rol } = req.body;
-  if (!email || !password || !usuario || !telefono || !rol)
-    return res.status(400).json({ error: "Faltan campos" });
-
-  try {
-    const hashed = await bcrypt.hash(password, 10);
-    db.run(
-      "INSERT INTO usuario (email, password_hash, usuario, telefono, rol) VALUES (?, ?, ?, ?, ?)",
-      [email, hashed, usuario, telefono, rol],
-      function (err) {
-        if (err) return res.status(400).json({ error: "Error al crear usuario" });
-        res.json({ id: this.lastID });
-      }
-    );
-  } catch {
-    res.status(500).json({ error: "Error del servidor" });
-  }
-});
-
-// Actualizar usuario
-app.put("/api/admin/users/:id", verifyAdmin, (req, res) => {
-  const { id } = req.params;
-  const { email, usuario, telefono, rol } = req.body;
-  db.run(
-    "UPDATE usuario SET email=?, usuario=?, telefono=?, rol=? WHERE id=?",
-    [email, usuario, telefono, rol, id],
-    function (err) {
-      if (err) return res.status(400).json({ error: "Error al actualizar" });
-      res.json({ cambios: this.changes });
-    }
-  );
-});
-
-// Eliminar usuario
-app.delete("/api/admin/users/:id", verifyAdmin, (req, res) => {
-  const { id } = req.params;
-  db.run("DELETE FROM usuario WHERE id=?", [id], function (err) {
-    if (err) return res.status(400).json({ error: "Error al eliminar" });
-    res.json({ eliminados: this.changes });
-  });
-});
-
-// Listar productos disponibles
-app.get("/api/productos", (req, res) => {
-  db.all("SELECT * FROM producto WHERE estado = 'disponible'", (err, rows) => {
-    if (err) return res.status(500).json({ error: "Error al obtener productos" });
-    res.json(rows);
-  });
-});
-
-// Crear producto
-app.post("/api/productos", (req, res) => {
-  const {
-    titulo, descripcion, precio, precio_anterior, descuento,
-    imagen, estado, resenas, calificacion, categoria
-  } = req.body;
-
-  if (!titulo || !precio)
-    return res.status(400).json({ error: "Faltan campos obligatorios (titulo, precio)" });
-
-  db.run(
-    `INSERT INTO producto (titulo, descripcion, precio, precio_anterior, descuento, imagen, estado, resenas, calificacion, categoria)
-     VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-    [
-      titulo, descripcion, precio, precio_anterior, descuento, imagen,
-      estado || "disponible", resenas || 0, calificacion || 5, categoria || "otros"
-    ],
-    function (err) {
-      if (err) return res.status(500).json({ error: "Error al insertar producto" });
-      res.json({ id: this.lastID });
-    }
-  );
-});
-
-// Actualizar producto
-app.put("/api/productos/:id", (req, res) => {
-  const { id } = req.params;
-  const {
-    titulo, descripcion, precio, precio_anterior, descuento,
-    imagen, estado, resenas, calificacion, categoria
-  } = req.body;
-
-  db.run(
-    `UPDATE producto SET 
-      titulo=?, descripcion=?, precio=?, precio_anterior=?, descuento=?, 
-      imagen=?, estado=?, resenas=?, calificacion=?, categoria=?
-     WHERE id=?`,
-    [
-      titulo, descripcion, precio, precio_anterior, descuento,
-      imagen, estado, resenas, calificacion, categoria, id
-    ],
-    function (err) {
-      if (err) return res.status(500).json({ error: "Error al actualizar producto" });
-      res.json({ success: true, cambios: this.changes });
-    }
-  );
-});
-
-// Eliminar producto
-app.delete("/api/productos/:id", (req, res) => {
-  const { id } = req.params;
-  db.run("DELETE FROM producto WHERE id = ?", [id], function (err) {
-    if (err) return res.status(500).json({ error: "Error al eliminar producto" });
-    res.json({ success: true, eliminados: this.changes });
-  });
-});
-
-// CREAR TRANSACCIÓN
-app.post("/api/crear-transaccion", async (req, res) => {
-  const { total } = req.body;
-
-  const buyOrder = `ORDEN${Math.floor(Math.random() * 1000000)}`;
-  const sessionId = `SESION${Math.floor(Math.random() * 1000000)}`;
-  const returnUrl = "http://localhost:5173/confirmar-transaccion";
-
-  try {
-    const response = await fetch(
-      "https://webpay3gint.transbank.cl/rswebpaytransaction/api/webpay/v1.2/transactions",
-      {
-        method: "POST",
-        headers: {
-          "Tbk-Api-Key-Id": "597055555532",
-          "Tbk-Api-Key-Secret": "579B532A7440BB0C9079DED94D31EA1615BACEB56610332264630D42D0A36B1C",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          buy_order: buyOrder,
-          session_id: sessionId,
-          amount: total,
-          return_url: returnUrl,
-        }),
-      }
-    );
-
-    const data = await response.json();
-    res.json({
-      url: `${data.url}?token_ws=${data.token}`,
-      token: data.token,
-    });
-  } catch (error) {
-    console.error("Error creando transacción:", error);
-    res.status(500).json({ error: error.message });
-  }
-});
-
-// CONFIRMAR TRANSACCIÓN
-app.post("/api/confirmar-transaccion", async (req, res) => {
-  const { token_ws } = req.body;
-
-  try {
-    const response = await fetch(
-      `https://webpay3gint.transbank.cl/rswebpaytransaction/api/webpay/v1.2/transactions/${token_ws}`,
-      {
-        method: "PUT",
-        headers: {
-          "Tbk-Api-Key-Id": "597055555532",
-          "Tbk-Api-Key-Secret": "579B532A7440BB0C9079DED94D31EA1615BACEB56610332264630D42D0A36B1C",
-          "Content-Type": "application/json",
-        },
-      }
-    );
-
-    const data = await response.json();
-
-    db.run(
-      `INSERT INTO orden (token, estado, monto, fecha) VALUES (?, ?, ?, datetime('now'))`,
-      [token_ws, data.status, data.amount],
-      function (err) {
-        if (err) console.error("❌ Error guardando orden:", err.message);
-        else console.log("✅ Orden guardada, ID:", this.lastID);
-      }
-    );
-
-    res.json(data);
-  } catch (error) {
-    console.error("Error confirmando transacción:", error);
-    res.status(500).json({ error: error.message });
-  }
-});
+// Rutas
+app.use("/api", require("./routes/auth"));
+app.use("/api", require("./routes/productos"));
+app.use("/api", require("./routes/usuarios"));
+app.use("/api", require("./routes/carrito")); // si tienes carrito
 
 // Iniciar servidor
 app.listen(PORT, () => {
